% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powGLM.R
\name{powGLM}
\alias{powGLM}
\title{Power Calculation for Generalized Linear Models}
\usage{
powGLM(n, alpha, delta, varZ, R2, f00)
}
\arguments{
\item{n}{Number of subjects}

\item{alpha}{Type I error rate}

\item{delta}{Vector of targeted effect size(s) for main variable(s). These correspond to values of the regression coefficients you want to detect.}

\item{varZ}{Variance for main variable(s); should be a matrix if more than one variable}

\item{R2}{Coefficient of determination matrix of main variable(s) given other covariate(s)}

\item{f00}{Expected residual variance under null, E[Var(Y|X)|H0]}
}
\value{
Power to detect targeted effect sizes with specified sample size and type I error rate
}
\description{
Calculate the power level attained for testing effects of main variable(s) in a generalized linear model (GLM)
}
\examples{
## Example of calculating power attained by 400 subjects for a 5\\% significance level test
num = 400
level = 0.05
del = c(0.5,0.3)
vz = diag(c(0.5,0.25))
rsq = diag(c(0,0.5))
f0 = 0.20
powGLM(num,level,del,vz,rsq,f0)
}
\references{
Martens, M.J. and Logan, B.L. (2020). A Unified Approach to Sample Size and Power Determination for Testing Parameters in Generalized Linear and Time to Event Regression Models. \emph{Statistics in Medicine} \strong{40(5)}, 1121-1132.

Martens, M.J., Kim, S., and Ahn, K.W. (2023+). Sample Size & Power Determination for Evaluating Multiple Parameters in Nonlinear Regression Models with Potential Stratification. \emph{Biometrics}, draft under review.
}
