% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspower.R
\docType{package}
\name{sspower}
\alias{sspower}
\title{Sample Size and Power Calculation for Regression Analysis}
\description{
Provides functions to compute sample size and power for studies that will be analyzed using generalized linear models and time-to-event regression models (Cox proportional hazards and Fine-Gray models).
}
\author{
Michael J. Martens \email{mmartens@mcw.edu}
}
