% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssTTE.R
\name{ssTTE}
\alias{ssTTE}
\title{Sample Size Calculation for Unstratified Time-to-Event Regression Models}
\usage{
ssTTE(alpha, power, delta, varZ, R2, sstype = "N", psi = NULL)
}
\arguments{
\item{alpha}{Type I error rate}

\item{power}{Power level}

\item{delta}{Vector of targeted effect size(s) for main variable(s). These correspond to values of the regression coefficients you want to detect.}

\item{varZ}{Variance for main variable(s); should be a matrix if more than one variable}

\item{R2}{Coefficient of determination matrix of main variable(s) given other covariate(s)}

\item{sstype}{Type of sample size calculation ('N' = patients (default), 'Events' = # events)}

\item{psi}{Event probability; required for patient-drive sample size (sstype = 'N')}
}
\value{
Required sample size to detect targeted effect sizes given specified type I error rate and power
}
\description{
Calculate the required sample size for testing effects of main variable(s) in unstratified time-to-event (Cox and Fine-Gray) regression models
}
\examples{
## Example of calculating sample size to attain 80\\% power for a 5\\% significance level test
level = 0.05
pow = 0.80
del = c(0.4,0.3)
vz = matrix(c(1,0.25,0.25,0.25),nrow=2)
rsq = matrix(c(0.5,0.25,0.25,0),nrow=2)
pr_event = 0.5
ssTTE(level,pow,del,vz,rsq,"Events")     # Event driven SS calculation
ssTTE(level,pow,del,vz,rsq,"N",pr_event) # Patient driven SS calculation
}
\references{
Martens, M.J. and Logan, B.L. (2020). A Unified Approach to Sample Size and Power Determination for Testing Parameters in Generalized Linear and Time to Event Regression Models. \emph{Statistics in Medicine} \strong{40(5)}, 1121-1132.

Martens, M.J., Kim, S., and Ahn, K.W. (2023+). Sample Size & Power Determination for Evaluating Multiple Parameters in Nonlinear Regression Models with Potential Stratification. \emph{Biometrics}, draft under review.
}
