% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssTTEstrat.R
\name{ssTTEstrat}
\alias{ssTTEstrat}
\title{Sample Size Calculation for Stratified Time-to-Event Regression Models}
\usage{
ssTTEstrat(alpha, power, delta, varZ, R2, psi, rho, sstype = "N")
}
\arguments{
\item{alpha}{Type I error rate}

\item{power}{Power level}

\item{delta}{Vector of targeted effect size(s) for main variable(s). These correspond to values of the regression coefficients you want to detect.}

\item{varZ}{3-dimensional array of variance matrices for main variable(s) in strata; each slice corresponds to a stratum}

\item{R2}{Array of coefficient of determination matrices of main variable(s) given other covariate(s) for strata; each slice corresponds to a stratum}

\item{psi}{Vector of event probabilities for strata}

\item{rho}{Vector of strata probabilities}

\item{sstype}{Type of sample size calculation ('N' = patients (default), 'Events' = # events)}
}
\value{
Required sample size to detect targeted effect sizes with specified type I error rate and power
}
\description{
Calculate the required sample size for testing effects of main variable(s) in stratified time-to-event (Cox and Fine-Gray) regression models
}
\examples{
## Example of calculating sample size to attain 80\\% power for a 5\\% significance level test
level = 0.05
pow = 0.80
del = c(0.4,0.3)
vz = array(0,c(2,2,2))
vz[,,1] = diag(rep(0.25,2))
vz[,,2] = diag(rep(1,2))
pr_event = c(0.2,0.3)
pr_strat = c(0.3,0.7)
rsq = array(0,c(2,2,2))
rsq[,,1] = diag(rep(0,2))
rsq[,,2] = matrix(c(0.5,0.25,0.25,0),nrow=2)
ssTTEstrat(level,pow,del,vz,rsq,pr_event,pr_strat,"Events")  # Event driven SS calculation
ssTTEstrat(level,pow,del,vz,rsq,pr_event,pr_strat,"N")       # Patient driven SS calculation
}
\references{
Martens, M.J. and Logan, B.L. (2020). A Unified Approach to Sample Size and Power Determination for Testing Parameters in Generalized Linear and Time to Event Regression Models. \emph{Statistics in Medicine} \strong{40(5)}, 1121-1132.

Martens, M.J., Kim, S., and Ahn, K.W. (2023+). Sample Size & Power Determination for Evaluating Multiple Parameters in Nonlinear Regression Models with Potential Stratification. \emph{Biometrics}, draft under review.
}
